+++ apis/dataset_metadata/dataset_date.py
@@ -1,5 +1,6 @@
 from typing import Any, Union
 
+from apis.authentication import is_granted
 from flask import request
 from flask_restx import Resource, fields
 
@@ -31,7 +32,10 @@ class DatasetDateResource(Resource):
     @api.doc("update date")
     @api.response(200, "Success")
     @api.response(400, "Validation Error")
-    def post(self, study_id: int, dataset_id: int):  # pylint: disable= unused-argument
+    def post(self, study_id: int, dataset_id: int):
+        study_obj = model.Study.query.get(study_id)
+        if not is_granted("dataset_metadata", study_obj):
+            return "Access denied, you can not make any change in dataset metadata", 403
         data: Union[Any, dict] = request.json
         data_obj = model.Dataset.query.get(dataset_id)
         list_of_elements = []
@@ -58,6 +62,9 @@ class DatasetDateDeleteResource(Resource):
     def delete(
         self, study_id: int, dataset_id: int, date_id: int
     ):  # pylint: disable= unused-argument
+        study_obj = model.Study.query.get(study_id)
+        if not is_granted("dataset_metadata", study_obj):
+            return "Access denied, you can not make any change in dataset metadata", 403
         date_ = model.DatasetDate.query.get(date_id)
 
         model.db.session.delete(date_)
