+++ apis/dataset_metadata/dataset_funder.py
@@ -1,5 +1,6 @@
 from typing import Any, Union
 
+from apis.authentication import is_granted
 from flask import request
 from flask_restx import Resource, fields
 
@@ -37,6 +38,9 @@ class DatasetFunderResource(Resource):
     @api.response(400, "Validation Error")
     def post(self, study_id: int, dataset_id: int):  # pylint: disable= unused-argument
         data: Union[Any, dict] = request.json
+        study_obj = model.Study.query.get(study_id)
+        if not is_granted("dataset_metadata", study_obj):
+            return "Access denied, you can not make any change in dataset metadata", 403
         data_obj = model.Dataset.query.get(dataset_id)
         list_of_elements = []
         for i in data:
@@ -61,10 +65,13 @@ class DatasetFunderUpdate(Resource):
     @api.response(400, "Validation Error")
     def delete(
         self,
-        study_id: int,  # pylint: disable= unused-argument
+        study_id: int,
         dataset_id: int,  # pylint: disable= unused-argument
         funder_id: int,
     ):
+        study_obj = model.Study.query.get(study_id)
+        if not is_granted("dataset_metadata", study_obj):
+            return "Access denied, you can not make any change in dataset metadata", 403
         dataset_funder_ = model.DatasetFunder.query.get(funder_id)
 
         model.db.session.delete(dataset_funder_)
